Implemented the functionality to load the circuit using the JSON, This feature is useful to load the circuit previous stored by user in JSON file.
Current we are using "gatesUIExtended.json" file as a sample file
to load the gates specified in it upon the click of "Load Circuit"
button.

The javascript file "loadjson.js" provides this functionality.

To achieve this minor refactoring also done in "cloneanddrag.js" and "gridtojson.js",
to avoid the code duplication.